Move level one commands
Moved to tracker module folder, like the other commands.
